+++ course-03/lesson-02/bootstrap/src/main/java/com/udacity/bootstrap/service/DogService.java
@@ -13,7 +13,7 @@ public interface DogService {
     List<String> retrieveDogBreeds();
 
     // retrieve a dog breed by its id:
-    String retrieveDogBreedById(int id);
+    String retrieveDogBreedById(Long id);
 
     // retrieve all dog names:
     List<String> retrieveDogNames();
